Warn instead of create an output without assets (#2037)
Fixes #1297

Since we don't know which assets will exist (only those which may exist)
before the build, save this check until we are outputting the directory.
The build that already happen may be a waste - but it will get reused if
the command is corrected to specify an output from a root that does have
assets in it.

- Move description of deleting contents to the specific place where we
  recognize that condition.
- Add a check for no outputs before creating the directory.
- Make it a severe log when the user chooses to not delete a conflicting
  directory.
- Tweak formatting of the prompt for shorter lines.